# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Teaching Planning Assistant API",
    "description": "Autonomous educational content generation engine for Chilean curriculum",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/content/adapt-content": {
      "post": {
        "tags": [
          "Content Generation"
        ],
        "summary": "Adapt Content",
        "description": "Adapt educational content for students with Special Educational Needs (NEE).\n\nTakes original content and adapts it based on student's NEE profile,\napplying evidence-based strategies for each identified need. Returns\nadapted content with teacher guidance and implementation checklist.\n\nSupported NEE types include: dyslexia, dyscalculia, ADHD, autism spectrum,\nvisual/hearing impairments, intellectual disabilities, and giftedness.\n\nAuthentication: JWT token OR API Key (with content:adapt permission)",
        "operationId": "adapt_content_api_v1_content_adapt_content_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdaptContentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdaptContentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/content/generate-quiz": {
      "post": {
        "tags": [
          "Content Generation"
        ],
        "summary": "Generate Quiz",
        "description": "Generate a quiz based on curriculum OAs.\n\nCreates a structured quiz with various question types suitable for\nformative or summative assessment. Questions are aligned to specific\nOAs and include Bloom's taxonomy levels.\n\nAuthentication: JWT token OR API Key (with content:generate permission)",
        "operationId": "generate_quiz_api_v1_content_generate_quiz_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateQuizRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateQuizResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/content/generate-activity": {
      "post": {
        "tags": [
          "Content Generation"
        ],
        "summary": "Generate Activity",
        "description": "Generate interactive learning activities based on curriculum OAs.\n\nCreates structured activities with steps, resources, timing,\nand differentiation options. Supports various activity types\nincluding group work, hands-on, games, and projects.",
        "operationId": "generate_activity_api_v1_content_generate_activity_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateActivityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateActivityResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/content/generate-exam": {
      "post": {
        "tags": [
          "Content Generation"
        ],
        "summary": "Generate Exam",
        "description": "Generate summative exams with multiple sections.\n\nCreates comprehensive exams with balanced coverage of OAs,\nvarious question types, and optional answer keys and rubrics.",
        "operationId": "generate_exam_api_v1_content_generate_exam_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateExamRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateExamResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/content/generate-reinforcement": {
      "post": {
        "tags": [
          "Content Generation"
        ],
        "summary": "Generate Reinforcement",
        "description": "Generate reinforcement materials for struggling students.\n\nCreates a personalized reinforcement plan with various materials\nlike concept reviews, practice exercises, visual guides, and\nworked examples. Includes guidance for teachers and parents.",
        "operationId": "generate_reinforcement_api_v1_content_generate_reinforcement_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateReinforcementRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateReinforcementResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/content/generate-summary": {
      "post": {
        "tags": [
          "Content Generation"
        ],
        "summary": "Generate Summary",
        "description": "Generate a unit summary for student study and review.\n\nCreates comprehensive summaries in various formats including\nbullet points, concept maps, and flashcard sets. Includes\nstudy questions, common misconceptions, and self-assessment.",
        "operationId": "generate_summary_api_v1_content_generate_summary_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateSummaryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateSummaryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdaptContentRequest": {
        "properties": {
          "content_type": {
            "$ref": "#/components/schemas/ContentType",
            "description": "Type of content"
          },
          "original_content": {
            "type": "string",
            "title": "Original Content",
            "description": "The original content to adapt (markdown or JSON string)"
          },
          "original_content_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Content Id",
            "description": "ID reference if from TPA"
          },
          "nee_types": {
            "items": {
              "$ref": "#/components/schemas/NEEType"
            },
            "type": "array",
            "title": "Nee Types",
            "description": "Student's NEE types"
          },
          "adaptation_level": {
            "$ref": "#/components/schemas/AdaptationLevel",
            "default": "moderate"
          },
          "student_grade_level": {
            "type": "string",
            "title": "Student Grade Level",
            "description": "Student's grade level"
          },
          "student_age": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Student Age",
            "description": "Student's age"
          },
          "student_strengths": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Student Strengths",
            "description": "Student's strengths"
          },
          "student_challenges": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Student Challenges",
            "description": "Specific challenges"
          },
          "previous_successful_strategies": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Successful Strategies",
            "description": "What has worked before"
          },
          "maintain_learning_objectives": {
            "type": "boolean",
            "title": "Maintain Learning Objectives",
            "description": "Keep same OAs",
            "default": true
          },
          "maintain_assessment_rigor": {
            "type": "boolean",
            "title": "Maintain Assessment Rigor",
            "description": "Same evaluation standard",
            "default": true
          },
          "include_teacher_guide": {
            "type": "boolean",
            "title": "Include Teacher Guide",
            "default": true
          },
          "include_visual_supports": {
            "type": "boolean",
            "title": "Include Visual Supports",
            "default": true
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "es"
          }
        },
        "type": "object",
        "required": [
          "content_type",
          "original_content",
          "nee_types",
          "student_grade_level"
        ],
        "title": "AdaptContentRequest",
        "description": "Request model for content adaptation"
      },
      "AdaptContentResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Whether adaptation succeeded"
          },
          "adapted_content": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AdaptedContent"
              },
              {
                "type": "null"
              }
            ],
            "description": "The adapted content"
          },
          "generation_metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Generation Metadata",
            "description": "Metadata about generation"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error",
            "description": "Error message if failed"
          }
        },
        "type": "object",
        "required": [
          "success",
          "generation_metadata"
        ],
        "title": "AdaptContentResponse",
        "description": "Response model for content adaptation"
      },
...
      "GenerateActivityResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Whether generation succeeded"
          },
          "activity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Activity"
              },
              {
                "type": "null"
              }
            ],
            "description": "The generated activity"
          },
          "generation_metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Generation Metadata",
            "description": "Metadata"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error",
            "description": "Error if failed"
          }
        },
        "type": "object",
        "required": [
          "success",
          "generation_metadata"
        ],
        "title": "GenerateActivityResponse"
      },
...
      "GenerateQuizRequest": {
        "properties": {
          "oa_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Oa Ids",
            "description": "OA identifiers to base quiz on"
          },
          "grade_level": {
            "type": "string",
            "title": "Grade Level",
            "description": "Grade level"
          },
          "subject": {
            "type": "string",
            "title": "Subject",
            "description": "Subject"
          },
          "num_questions": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Num Questions",
            "description": "Number of questions",
            "default": 10
          },
          "question_types": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/QuestionType"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question Types",
            "description": "Allowed question types (default: all)"
          },
          "difficulty_distribution": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Difficulty Distribution",
            "description": "Distribution like {'easy': 3, 'medium': 5, 'hard': 2}"
          },
          "bloom_levels": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/BloomLevel"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bloom Levels",
            "description": "Target Bloom's levels"
          },
          "time_limit_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Limit Minutes",
            "description": "Time limit"
          },
          "context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context",
            "description": "Additional context"
          },
          "language": {
            "type": "string",
            "title": "Language",
            "description": "Output language",
            "default": "es"
          },
          "include_hints": {
            "type": "boolean",
            "title": "Include Hints",
            "description": "Include hints",
            "default": true
          },
          "include_explanations": {
            "type": "boolean",
            "title": "Include Explanations",
            "description": "Include explanations",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "oa_ids",
          "grade_level",
          "subject"
        ],
        "title": "GenerateQuizRequest",
        "description": "Request model for quiz generation"
      },
      "GenerateQuizResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Whether generation succeeded"
          },
          "quiz": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Quiz-Output"
              },
              {
                "type": "null"
              }
            ],
            "description": "The generated quiz"
          },
          "generation_metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Generation Metadata",
            "description": "Metadata about generation"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error",
            "description": "Error message if failed"
          }
        },
        "type": "object",
        "required": [
          "success",
          "generation_metadata"
        ],
        "title": "GenerateQuizResponse",
        "description": "Response model for quiz generation"
      },
...
      "GenerateSummaryResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "summary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UnitSummary-Output"
              },
              {
                "type": "null"
              }
            ]
          },
          "generation_metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Generation Metadata"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "success",
          "generation_metadata"
        ],
        "title": "GenerateSummaryResponse",
        "description": "Response model for summary generation"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "token"
          }
        }
      },
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}